+++ src/main/scala/chisel3/util/MuxImpl.scala
@@ -7,8 +7,6 @@ package chisel3.util
 
 import chisel3._
 import chisel3.experimental.SourceInfo
-import chisel3.internal.sourceinfo.MuxLookupTransform
-import scala.language.experimental.macros
 
 /** Builds a Mux tree out of the input signal vector using a one hot encoded
   * select signal. Returns the output of the Mux tree.
@@ -52,43 +50,24 @@ object PriorityMux {
   def apply[T <: Data](sel: Bits, in: Seq[T]): T = apply((0 until in.size).map(sel(_)), in)
 }
 
-/** Creates a cascade of n Muxs to search for a key value. The Selector may be a UInt or an EnumType.
-  *
-  * @example {{{
-  * MuxLookup(idx, default)(Seq(0.U -> a, 1.U -> b))
-  * MuxLookup(myEnum, default)(Seq(MyEnum.a -> 1.U, MyEnum.b -> 2.U, MyEnum.c -> 3.U))
-  * }}}
-  */
-object MuxLookup extends SourceInfoDoc {
+private[chisel3] trait MuxLookupImpl {
 
-  /** @param key a key to search for
-    * @param default a default value if nothing is found
-    * @param mapping a sequence to search of keys and values
-    * @return the value found or the default if not
-    */
-  def apply[T <: Data](key: UInt, default: T)(mapping: Seq[(UInt, T)]): T =
-    macro MuxLookupTransform.applyCurried[UInt, T]
-
-  /** @param key a key to search for
-    * @param default a default value if nothing is found
-    * @param mapping a sequence to search of keys and values
-    * @return the value found or the default if not
-    */
-  def apply[S <: EnumType, T <: Data](key: S, default: T)(mapping: Seq[(S, T)]): T =
-    macro MuxLookupTransform.applyEnum[S, T]
-
-  /** @group SourceInfoTransformMacro */
-  def do_applyEnum[S <: EnumType, T <: Data](
+  protected def _applyEnumImpl[S <: EnumType, T <: Data](
     key:     S,
     default: T,
     mapping: Seq[(S, T)]
   )(
     implicit sourceinfo: SourceInfo
   ): T =
-    do_apply[UInt, T](key.asUInt, default, mapping.map { case (s, t) => (s.asUInt, t) })
+    _applyImpl[UInt, T](key.asUInt, default, mapping.map { case (s, t) => (s.asUInt, t) })
 
-  /** @group SourceInfoTransformMacro */
-  def do_apply[S <: UInt, T <: Data](key: S, default: T, mapping: Seq[(S, T)])(implicit sourceinfo: SourceInfo): T = {
+  protected def _applyImpl[S <: UInt, T <: Data](
+    key:     S,
+    default: T,
+    mapping: Seq[(S, T)]
+  )(
+    implicit sourceinfo: SourceInfo
+  ): T = {
     /* If the mapping is defined for all possible values of the key, then don't use the default value */
     val (defaultx, mappingx) = key.widthOption match {
       case Some(width) =>
